+++ src/mrc_mat_mcsr.c
@@ -63,7 +63,7 @@ mrc_mat_mcsr_destroy(struct mrc_mat *mat)
 // mrc_mat_mcsr_add_value
 
 static void
-mrc_mat_mcsr_add_value(struct mrc_mat *mat, int row_idx, int col_idx, float val)
+mrc_mat_mcsr_add_value(struct mrc_mat *mat, int row_idx, int col_idx, double val)
 {
   struct mrc_mat_mcsr *sub = mrc_mat_mcsr(mat);
